SentimentAnalysisSTock/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-socketio==5.3.6
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
nltk==3.8.1
yfinance==0.2.28
praw==7.7.1
newsapi-python==0.2.7
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
plotly==5.17.0
python-dateutil==2.8.2